worked on adding

+++ src/emulator/BitOperations.h
@@ -1,6 +1,6 @@
 #include <iostream>
 
-int add(int a, int b);
+uint8_t add(uint8_t a, uint8_t b);
 
 int mul(int  a, int b);
